Move useTransactionTokens hook
