+++ apps/desktop-wallet/src/features/transactionsDisplay/transactionDetailsModal/FTsDataListRow.tsx
@@ -19,10 +19,10 @@ along with the library. If not, see <http://www.gnu.org/licenses/>.
 import { useTranslation } from 'react-i18next'
 import styled from 'styled-components'
 
-import useTransactionTokens from '@/api/apiDataHooks/useTransactionTokens'
 import Amount from '@/components/Amount'
 import DataList from '@/components/DataList'
 import { TransactionDetailsModalSectionProps } from '@/features/transactionsDisplay/transactionDetailsModal/types'
+import useTransactionTokens from '@/features/transactionsDisplay/useTransactionTokens'
 import { useTransactionInfoType } from '@/utils/transactions'
 
 const FTsDataListRow = ({ tx, addressHash }: TransactionDetailsModalSectionProps) => {
